\name{flips.demo}
\alias{flips.demo}

\title{FLIPS demonstration}

\description{A simple FLIPS example}

\usage{flips.demo()}

\value{None.}

\details{This is a simple FLIPS example of noisy deconvolution
problem. The effect of regularization is demonstrated along with the
marginalization of unknowns.

The main reason of the existance of this is that it is easy to check
that the FLIPS works properly.}

\author{Mikko Orispaa (University of Oulu, Finland)
\email{mikko.orispaa@oulu.fi}}

\keyword{algebra}